Rewrite Task 4.
