--- JahorLaryn/Task4.2.1.py
@@ -1,10 +1,0 @@
-a = "I am global variable!"
-
-def enclosing_funcion():
-    a = "I am variable from enclosed function!"
-    def inner_function():
-        a = "I am local variable!"
-        print(a)
-    inner_function()
-
-enclosing_funcion()
