--- JahorLaryn/Task4.2.2.py
@@ -1,7 +1,0 @@
-a = "I am global variable!"
-
-def enclosing_funcion():
-    a = "I am variable from enclosed function!"
-    def inner_function():
-        a = "I am local variable!"
-        print(a)
